开发维修工单列表，完善创建故障，创建设备功能，设备选择功能

+++ android/FactoryRP/app/src/main/java/tech/yozo/factoryrp/vo/req/AddDeviceReq.java
@@ -60,7 +60,7 @@ public class AddDeviceReq implements Serializable{
     private String head;
 
 //    @ApiModelProperty(value = "使用状况",notes ="使用状况" )
-    private Integer useStatus;
+    private Long useStatus;
 
 //    @ApiModelProperty(value = "安装地点",notes ="安装地点" )
     private String installationAddress;
